refactor(components): reorganize auth, layout and route components
